refactor: view 분리
PurchaseSectionView를 아래 두 가지로 분리 및 변경
- ChargeSubmitFormView
- PurchasedTicketListSectionView

WinningResultSectionView를 아래 두 가지로 분리 및 변경
- WinningNumberSubmitFormView
- WinningResultModalView
