+++ src/js/views/LottoMachineView.js
@@ -1,33 +1,45 @@
-import WinningResultSectionView from './WinningResultSectionView';
-import PurchaseTicketSectionView from './PurchaseTicketSectionView';
 import { $ } from '../utils/util';
+
+import ChargeSubmitFormView from './ChargeSubmitFormView';
+import PurchasedTicketListSectionView from './PurchasedTicketListSectionView';
+import WinningResultModalView from './WinningResultModalView';
+import WinningNumberSubmitFormView from './WinningNumberSubmitFormView';
 
 export default class LottoMachineView {
   #app;
 
-  #purchaseTicketSectionView;
+  #chargeSubmitFormView;
 
-  #winningResultSectionView;
+  #purchasedTicketListSectionView;
+
+  #winningNumberSubmitFormView;
+
+  #winningResultModalView;
 
   constructor() {
     this.#app = $('#app');
-    this.#purchaseTicketSectionView = new PurchaseTicketSectionView(this.#app);
-    this.#winningResultSectionView = new WinningResultSectionView(this.#app);
+    this.#chargeSubmitFormView = new ChargeSubmitFormView(this.#app);
+    this.#purchasedTicketListSectionView = new PurchasedTicketListSectionView(this.#app);
+    this.#winningNumberSubmitFormView = new WinningNumberSubmitFormView(this.#app);
+    this.#winningResultModalView = new WinningResultModalView(this.#app);
   }
 
   get app() { return this.#app; }
 
   initialize(lottos) {
-    this.#purchaseTicketSectionView.initialize(lottos);
-    this.#winningResultSectionView.initialize();
+    this.#chargeSubmitFormView.initialize();
+    this.#purchasedTicketListSectionView.initialize(lottos);
+    this.#winningNumberSubmitFormView.initialize();
+    this.#winningResultModalView.initialize();
   }
 
   updateOnPurchase(tickets, charge) {
-    this.#purchaseTicketSectionView.updateOnPurchase(tickets, charge);
-    this.#winningResultSectionView.updateOnPurchase(tickets);
+    this.#chargeSubmitFormView.updateOnPurchase(tickets, charge);
+    this.#purchasedTicketListSectionView.updateOnPurchase(tickets);
+    this.#winningNumberSubmitFormView.updateOnPurchase(tickets);
   }
 
   updateOnCheckWinningResult(winningResult) {
-    this.#winningResultSectionView.updateOnCheckWinningResult(winningResult);
+    this.#winningResultModalView.updateOnCheckWinningResult(winningResult);
   }
 }
